+++ Plugins/TableProPluginKit/TableProPlugin.swift
@@ -5,6 +5,11 @@ public protocol TableProPlugin: AnyObject {
     static var pluginVersion: String { get }
     static var pluginDescription: String { get }
     static var capabilities: [PluginCapability] { get }
+    static var dependencies: [String] { get }
 
     init()
+}
+
+public extension TableProPlugin {
+    static var dependencies: [String] { [] }
 }